:building_construction: Moved AppointmentController store to service
all changes are functional
